requirements.txt
langchain
langchain-community
langchain-ollama
langchain_groq
langchain-docling
langchain-tavily
langgraph
transformers
PyMuPDF
pdfplumber
python-pptx
torch
Pillow
pydantic
fastapi
uvicorn
python-dotenv
pandas
numpy
faiss-cpu
python-multipart
aiofiles
websockets
selenium
mcp[cli]
ipykernel
